assignment 7

+++ src/main/java/com/product/crud/Controller/StorageController.java
@@ -7,7 +7,10 @@ import com.product.crud.model.Product;
 import com.product.crud.model.ResponseObject;
 import com.product.crud.services.ProductService;
 import com.product.crud.services.StorageService;
+import com.timgroup.statsd.StatsDClient;
 import jakarta.servlet.http.HttpServletRequest;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.HttpStatus;
@@ -25,10 +28,15 @@ public class StorageController {
     @Autowired
     private ProductService productService;
 
+    @Autowired
+    private StatsDClient statsDClient;
+    Logger log = LoggerFactory.getLogger(StorageController.class);
+
     @RequestMapping(path = "/v1/product/{product_id}/image", method = RequestMethod.POST)
     @ResponseBody
     public ResponseEntity<?> uploadFile( @PathVariable("product_id") Integer product_id,@RequestParam(value="file")MultipartFile file, HttpServletRequest request){
-
+        log.info("Inside Storage Controller. Uploading Image");
+        statsDClient.incrementCounter("endpoint.uploadFile.http.post");
         try {
             if (!(productService.isAuthorisedForPut(product_id, request.getHeader("Authorization").split(" ")[1], null))) {
                 throw new InvalidInputException("Invalid Username or Password");
@@ -47,6 +55,11 @@ public class StorageController {
             response.setHttpStatusCode(HttpStatus.FORBIDDEN);
             response.setResponseMessage(e.getMessage());
             return new ResponseEntity<ResponseObject>(response,HttpStatus.FORBIDDEN);
+        } catch (Exception e) {
+            ResponseObject response = new ResponseObject();
+            response.setHttpStatusCode(HttpStatus.BAD_REQUEST);
+            response.setResponseMessage("Error Occured while Uploading. Please make sure the file is not empty or currupt" + e.getMessage());
+            return new ResponseEntity<ResponseObject>(response,HttpStatus.FORBIDDEN);
         }
 
     }
@@ -54,6 +67,8 @@ public class StorageController {
     @RequestMapping(path = "/v1/product/download/{imageName}", method = RequestMethod.GET)
     public ResponseEntity<?> downloadImage(@PathVariable String imageName){
 
+        log.info("Inside Storage Controller. Getting All Images");
+        statsDClient.incrementCounter("endpoint.downloadImage.http.get");
 
         byte[] data = storageService.downloadImage(imageName);
         ByteArrayResource resource = new ByteArrayResource(data);
@@ -66,9 +81,19 @@ public class StorageController {
 
     @RequestMapping(path = "/v1/product/{product_id}/image/{image_id}", method = RequestMethod.GET)
     public ResponseEntity<?> getImageDetails(@PathVariable Integer product_id,@PathVariable Integer image_id,HttpServletRequest request){
+        log.info("Inside Storage Controller. Getting Image");
+        statsDClient.incrementCounter("endpoint.downloadImage.http.get");
         try {
+            if(productService.findProductById(product_id)<1){
+                throw new InvalidInputException("Invalid Product ID");
+            }
+
             if (!(productService.isAuthorisedForPut(product_id, request.getHeader("Authorization").split(" ")[1], null))) {
                 throw new InvalidInputException("Invalid Username or Password");
+            }
+            //Authorise Product with ID
+            if(!storageService.isAuthorisedWithProduct(image_id,product_id)){
+                throw new InvalidInputException("Invalid Product ID for the Image");
             }
             Image image = storageService.getImage(image_id);
             if(image!=null){
@@ -82,6 +107,7 @@ public class StorageController {
             response.setHttpStatusCode(HttpStatus.BAD_REQUEST);
             response.setResponseMessage(e.getMessage());
             if(e.getMessage().matches("Invalid Username or Password")) return new ResponseEntity<ResponseObject>(response,HttpStatus.UNAUTHORIZED);
+            if(e.getMessage().matches("Invalid Product ID")) return new ResponseEntity<ResponseObject>(response,HttpStatus.NOT_FOUND);
             return new ResponseEntity<ResponseObject>(response,HttpStatus.BAD_REQUEST);
         }catch(UserAuthorizationException e){
             ResponseObject response = new ResponseObject();
@@ -93,6 +119,8 @@ public class StorageController {
 
     @RequestMapping(path = "/v1/product/{product_id}/image", method = RequestMethod.GET)
     public ResponseEntity<?> getImageList(@PathVariable Integer product_id,HttpServletRequest request){
+        log.info("Inside Storage Controller. Getting Image");
+        statsDClient.incrementCounter("endpoint.downloadImage.http.get");
         try {
             if (!(productService.isAuthorisedForPut(product_id, request.getHeader("Authorization").split(" ")[1], null))) {
                 throw new InvalidInputException("Invalid Username or Password");
@@ -124,9 +152,17 @@ public class StorageController {
     @RequestMapping(path = "/v1/product/{product_id}/image/{image_id}", method = RequestMethod.DELETE)
     @ResponseBody
     public ResponseEntity<?> deleteImage(@PathVariable Integer image_id,@PathVariable Integer product_id,HttpServletRequest request){
+        log.info("Inside Storage Controller. Getting Image");
+        statsDClient.incrementCounter("endpoint.downloadImage.http.get");
         try {
+            if(productService.findProductById(product_id)<1){
+                throw new InvalidInputException("Invalid Product ID");
+            }
             if (!(productService.isAuthorisedForPut(product_id, request.getHeader("Authorization").split(" ")[1], null))) {
                 throw new InvalidInputException("Invalid Username or Password");
+            }
+            if(!storageService.isAuthorisedWithProduct(image_id,product_id)){
+                throw new InvalidInputException("Invalid Product ID for the Image");
             }
             String deleteResponse=storageService.deleteImage(image_id);
             if(deleteResponse.equals("Image Not Found"))  return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
@@ -136,6 +172,7 @@ public class StorageController {
             response.setHttpStatusCode(HttpStatus.BAD_REQUEST);
             response.setResponseMessage(e.getMessage());
             if(e.getMessage().matches("Invalid Username or Password")) return new ResponseEntity<ResponseObject>(response,HttpStatus.UNAUTHORIZED);
+            if(e.getMessage().matches("Invalid Product ID")) return new ResponseEntity<ResponseObject>(response,HttpStatus.NOT_FOUND);
             return new ResponseEntity<ResponseObject>(response,HttpStatus.BAD_REQUEST);
         }catch(UserAuthorizationException e){
             ResponseObject response = new ResponseObject();
